Prelaunch added to devrel-tools
Posting the Prelaunch script to dev-rel tools. It is now updated to automatically automatically pull the developers who have updated their account internally (and are in the "Pending" state) so that the entire process of generating Jira tickets is automated when you run the script.

Before running, please fetch the folder from GitHub to make sure the last date ran is actually the most recent date the script was run.