Override of Autocredit now available
Autocredit can now be overridden in terminal if you select option 4 instead of option 1. This is useful for when, say, something happened to your virtualenv and in the process of fixing it you found that the timestamp updated without adding the requisite Jira tickets. Now you can manually enter those Jira tickets to avoid silly fiascos like this one.